LPD-51750 Adapt unit tests to refactors
