+++ modules/apps/frontend-js/frontend-js-web/test/liferay/util/form/set_form_values.es.js
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import buildFragment from '../../../../src/main/resources/META-INF/resources/liferay/util/build_fragment';
 import getFormElement from '../../../../src/main/resources/META-INF/resources/liferay/util/form/get_form_element.es';
 import setFormValues from '../../../../src/main/resources/META-INF/resources/liferay/util/form/set_form_values.es';
+import buildFragment from '../../../../src/main/resources/META-INF/resources/main/util/build_fragment';
 
 describe('Liferay.Util.setFormValues', () => {
 	it('sets the given values of form elements', () => {
