+++ modules/apps/frontend-js/frontend-js-web/test/liferay/util/storage.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {CONSENT_TYPES} from '../../../src/main/resources/META-INF/resources/liferay/util/consent';
+import {CONSENT_TYPES} from '../../../src/main/resources/META-INF/resources/liferay/util/cookie/cookie';
 import Storage from '../../../src/main/resources/META-INF/resources/liferay/util/storage';
 
 const anyName = 'any-name';
