+++ modules/apps/frontend-js/frontend-js-web/test/main/utils/date_time/parse.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {parse} from '../../../src/main/resources/META-INF/resources/utils/date_time/parse';
+import {parse} from '../../../../src/main/resources/META-INF/resources/main/utils/date_time/parse';
 
 describe('parse', () => {
 	it('parses date string "01/01/2025" as Date object', () => {
